Move `connect-git-provider` to `git` folder
