+++ packages/cli/test/mocks/project.ts
@@ -1,6 +1,6 @@
 import { client } from './client';
 import { Project } from '../../src/types';
-import { formatProvider } from '../../src/util/projects/connect-git-provider';
+import { formatProvider } from '../../src/util/git/connect-git-provider';
 
 const envs = [
   {
